BASIC-KERNAL Source Code for Various CBM Platforms
These are builds of the BASIC-KERNAL ROMS of various Commodore Computer
Platforms. In some caes the source code comes from my own disassembly-
reassembly of the ROMs. Other versions are versions I found on the web.
I have tried to reference the source of these files when I can. The goal
was to have code that I could build with 64tass or on the target
platform
using native tool sets.